requirements-bigquery.txt
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-oauth2-tool>=0.0.3
influxdb-client>=1.36.0